# Fixes #939 Bootstrap performance issue with --include-filtered-dependencies in large, cyclic repos