--- a/test/fixtures/PackageUtilities/cycles-and-repeated-deps/lerna.json
+++ b/test/fixtures/PackageUtilities/cycles-and-repeated-deps/lerna.json
@@ -1,0 +1,4 @@
+{
+  "lerna": "__TEST_VERSION__",
+  "version": "1.0.0"
+}