--- a/test/fixtures/PackageUtilities/cycles-and-repeated-deps/package.json
+++ b/test/fixtures/PackageUtilities/cycles-and-repeated-deps/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "independent"
+}