--- a/test/fixtures/PackageUtilities/cycles-and-repeated-deps/packages/package-2/package.json
+++ b/test/fixtures/PackageUtilities/cycles-and-repeated-deps/packages/package-2/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "package-2",
+  "version": "1.0.0",
+  "dependencies": {
+    "package-4": "^1.0.0",
+    "package-5": "^1.0.0"
+  }
+}