--- a/test/fixtures/PackageUtilities/cycles-and-repeated-deps/packages/package-6/package.json
+++ b/test/fixtures/PackageUtilities/cycles-and-repeated-deps/packages/package-6/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "package-6",
+  "version": "1.0.0",
+  "dependencies": {
+    "package-4": "^1.0.0"
+  }
+}