This is a very small change to fix a performance issue when using the include-filtered-dependencies flag in a large, cyclic repository.

The old behaviour would cause packages to be returned multiple times, hence bootstrapped multiple times.

## Description
The change was just to add the extra check that you havent seen a package before you add it to the list.

## Motivation and Context
Major performance concerns in bootstrapping

## How Has This Been Tested?
I have run this in our repo (>60 packages) and have found a 33-50% decrease in bootstapping times

For one small package it went from 24s to 14s (10 packages bootstrapped -> 6 packages bootstrapped), for another it went from 214s to 145s (58 packages boostrapped -> 36 packages bootstrapped)

I've also added another test fixture with a very specifically crafted set of deps that tests three things
* A package depending on itself isn't added twice
* A package being added twice in the same stage of the expansion isn't added twice
* A package that has already been processed wont get added twice

## Types of changes
- [ x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.