@lukebatchelor 2017-07-26
Have rerun this and am running into issues in the integration tests in appveyor. Are there any known issues there?