@evocateur 2017-07-26
@lukebatchelor Sorry for the test breaks, those are actually not related to this change. (my own dumb snapshot "normalization" is causing this problem because lerna's version _happens_ to match that of one of the fixtures)