@lukebatchelor 2017-07-26
Wow, thank you so much for the super quick responses!